+++ src/test/kotlin/domaintest/RefereeTest.kt
@@ -1,4 +1,4 @@
-package domain_test
+package domaintest
 
 import domain.game.CarRacingGameReferee
 import org.assertj.core.api.Java6Assertions.assertThat
